+++ src/main/java/com/spring/sharepod/v1/controller/NoticeRestController.java
@@ -3,8 +3,8 @@ package com.spring.sharepod.v1.controller;
 
 import com.spring.sharepod.entity.User;
 import com.spring.sharepod.v1.dto.response.BasicResponseDTO;
-import com.spring.sharepod.v1.dto.response.NoticeCount;
-import com.spring.sharepod.v1.dto.response.NoticeInfoResponseDto;
+import com.spring.sharepod.v1.dto.response.notice.NoticeCount;
+import com.spring.sharepod.v1.dto.response.notice.NoticeInfoResponseDto;
 import com.spring.sharepod.v1.service.NoticeService;
 import com.spring.sharepod.v1.validator.TokenValidator;
 import lombok.RequiredArgsConstructor;
