+++ src/main/java/com/spring/sharepod/v1/controller/UserRestController.java
@@ -3,9 +3,7 @@ package com.spring.sharepod.v1.controller;
 import com.spring.sharepod.entity.User;
 import com.spring.sharepod.v1.dto.request.UserRequestDto;
 import com.spring.sharepod.v1.dto.response.*;
-import com.spring.sharepod.v1.dto.response.User.UserInfoResponseDto;
-import com.spring.sharepod.v1.dto.response.User.UserReservation;
-import com.spring.sharepod.v1.dto.response.User.UserResponseDto;
+import com.spring.sharepod.v1.dto.response.User.*;
 import com.spring.sharepod.v1.service.AwsS3Service;
 import com.spring.sharepod.v1.service.UserService;
 import com.spring.sharepod.v1.validator.TokenValidator;
