+++ src/main/java/com/spring/sharepod/v1/dto/response/Auth/AuthBool.java
@@ -1,5 +1,4 @@
-package com.spring.sharepod.v1.dto.response;
-
+package com.spring.sharepod.v1.dto.response.Auth;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -8,8 +7,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 @Data
-public class NoticeCount {
+public class AuthBool {
     private String result;
     private String msg;
-    private int noticeCnt;
 }
