--- src/main/java/com/spring/sharepod/v1/dto/response/AuthBool.java
@@ -1,13 +1,0 @@
-package com.spring.sharepod.v1.dto.response;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@AllArgsConstructor
-@NoArgsConstructor
-@Data
-public class AuthBool {
-    private String result;
-    private String msg;
-}
