+++ src/main/java/com/spring/sharepod/v1/dto/response/Board/BoardList.java
@@ -1,4 +1,4 @@
-package com.spring.sharepod.v1.dto.response;
+package com.spring.sharepod.v1.dto.response.Board;
 
 import com.spring.sharepod.v1.dto.response.Board.BoardAllResponseDto;
 import lombok.AllArgsConstructor;
