+++ src/main/java/com/spring/sharepod/v1/dto/response/User/UserInfo.java
@@ -1,16 +1,14 @@
-package com.spring.sharepod.v1.dto.response;
+package com.spring.sharepod.v1.dto.response.User;
 
 import com.spring.sharepod.v1.dto.response.User.UserInfoResponseDto;
 import lombok.AllArgsConstructor;
-import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
 @AllArgsConstructor
 @NoArgsConstructor
 @Data
-@Builder
-public class UserMyInfoResponseDto {
+public class UserInfo {
     private String result;
     private String msg;
     private UserInfoResponseDto userInfo;
