+++ src/main/java/com/spring/sharepod/v1/dto/response/User/UserMyInfoResponseDto.java
@@ -1,20 +1,18 @@
-package com.spring.sharepod.v1.dto.response;
+package com.spring.sharepod.v1.dto.response.User;
 
-import com.spring.sharepod.v1.dto.response.notice.Notice;
+import com.spring.sharepod.v1.dto.response.User.UserInfoResponseDto;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.util.List;
-
 @AllArgsConstructor
 @NoArgsConstructor
 @Data
 @Builder
-public class NoticeInfoResponseDto {
+public class UserMyInfoResponseDto {
     private String result;
     private String msg;
-    private List<Notice> noticeList;
+    private UserInfoResponseDto userInfo;
 
 }
