+++ src/main/java/com/spring/sharepod/v1/dto/response/User/UserOrder.java
@@ -1,8 +1,8 @@
-package com.spring.sharepod.v1.dto.response;
+package com.spring.sharepod.v1.dto.response.User;
 
+import com.spring.sharepod.v1.dto.response.*;
 import com.spring.sharepod.v1.dto.response.User.UserReservation;
 import lombok.AllArgsConstructor;
-import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
@@ -11,8 +11,7 @@ import java.util.List;
 @AllArgsConstructor
 @NoArgsConstructor
 @Data
-@Builder
-public class UserOrderResponseDto {
+public class UserOrder {
     private String result;
     private String msg;
     private List<RentBuyer> rentBuyerList;
