+++ src/main/java/com/spring/sharepod/v1/dto/response/User/UserOrderResponseDto.java
@@ -1,8 +1,10 @@
-package com.spring.sharepod.v1.dto.response;
+package com.spring.sharepod.v1.dto.response.User;
 
-import com.spring.sharepod.v1.dto.response.*;
+import com.spring.sharepod.v1.dto.response.RentBuyer;
+import com.spring.sharepod.v1.dto.response.RentSeller;
 import com.spring.sharepod.v1.dto.response.User.UserReservation;
 import lombok.AllArgsConstructor;
+import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
@@ -11,7 +13,8 @@ import java.util.List;
 @AllArgsConstructor
 @NoArgsConstructor
 @Data
-public class UserOrder {
+@Builder
+public class UserOrderResponseDto {
     private String result;
     private String msg;
     private List<RentBuyer> rentBuyerList;
