+++ src/main/java/com/spring/sharepod/v1/dto/response/notice/NoticeCount.java
@@ -1,6 +1,6 @@
-package com.spring.sharepod.v1.dto.response;
+package com.spring.sharepod.v1.dto.response.notice;
 
-import com.spring.sharepod.v1.dto.response.User.UserInfoResponseDto;
+
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -8,9 +8,8 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 @Data
-public class UserInfo {
+public class NoticeCount {
     private String result;
     private String msg;
-    private UserInfoResponseDto userInfo;
-
+    private int noticeCnt;
 }
