+++ src/main/java/com/spring/sharepod/v1/dto/response/notice/NoticeInfoResponseDto.java
@@ -1,0 +1,20 @@
+package com.spring.sharepod.v1.dto.response.notice;
+
+import com.spring.sharepod.v1.dto.response.notice.Notice;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import java.util.List;
+
+@AllArgsConstructor
+@NoArgsConstructor
+@Data
+@Builder
+public class NoticeInfoResponseDto {
+    private String result;
+    private String msg;
+    private List<Notice> noticeList;
+
+}
